move hepmc3 tests to separate file
