+++ Examples/Python/tests/test_writer.py
@@ -442,88 +442,6 @@ def test_csv_multitrajectory_writer(tmp_path):
     assert all(f.stat().st_size > 20 for f in csv_dir.iterdir())
 
 
-@pytest.fixture(scope="session")
-def hepmc_data_impl(tmp_path_factory):
-    import subprocess
-
-    script = (
-        Path(__file__).parent.parent.parent.parent
-        / "Examples"
-        / "Scripts"
-        / "Python"
-        / "event_recording.py"
-    )
-    assert script.exists()
-
-    with tempfile.TemporaryDirectory() as tmp_path:
-        env = os.environ.copy()
-        env["NEVENTS"] = "1"
-        subprocess.check_call([sys.executable, str(script)], cwd=tmp_path, env=env)
-
-        outfile = Path(tmp_path) / "hepmc3/event000000000-events.hepmc3"
-        # fake = Path("/scratch/pagessin/acts/hepmc3/event000000000-events.hepmc3")
-
-        # outfile.parent.mkdir()
-        # shutil.copy(fake, outfile)
-
-        assert outfile.exists()
-
-        yield outfile
-
-
-@pytest.fixture
-def hepmc_data(hepmc_data_impl: Path, tmp_path):
-    dest = tmp_path / hepmc_data_impl.name
-    shutil.copy(hepmc_data_impl, dest)
-
-    return dest
-
-
-@pytest.mark.skipif(not hepmc3Enabled, reason="HepMC3 plugin not available")
-@pytest.mark.skipif(not dd4hepEnabled, reason="DD4hep not set up")
-@pytest.mark.skipif(not geant4Enabled, reason="Geant4 not set up")
-@pytest.mark.odd
-@pytest.mark.slow
-def test_hepmc3_histogram(hepmc_data, tmp_path):
-    from acts.examples.hepmc3 import (
-        HepMC3AsciiReader,
-        HepMCProcessExtractor,
-    )
-
-    s = Sequencer(numThreads=1)
-
-    s.addReader(
-        HepMC3AsciiReader(
-            level=acts.logging.INFO,
-            inputDir=str(hepmc_data.parent),
-            inputStem="events",
-            outputEvents="hepmc-events",
-        )
-    )
-
-    s.addAlgorithm(
-        HepMCProcessExtractor(
-            level=acts.logging.INFO,
-            inputEvents="hepmc-events",
-            extractionProcess="Inelastic",
-        )
-    )
-
-    # This segfaults, see https://github.com/acts-project/acts/issues/914
-    # s.addWriter(
-    #     RootNuclearInteractionParametersWriter(
-    #         level=acts.logging.INFO, inputSimulationProcesses="event-fraction"
-    #     )
-    # )
-
-    alg = AssertCollectionExistsAlg(
-        "hepmc-events", name="check_alg", level=acts.logging.INFO
-    )
-    s.addAlgorithm(alg)
-
-    s.run()
-
-
 @pytest.mark.edm4hep
 @pytest.mark.skipif(not edm4hepEnabled, reason="EDM4hep is not set up")
 def test_edm4hep_measurement_writer(tmp_path, fatras):
